Started refactoring into multiple files. ST seems to be working again.
Separated the components into a directory and different files.
