+++ fsleyes_plugin_shimming_toolbox/__init__.py
@@ -1,0 +1,3 @@
+import os
+
+__dir_st_plugin__ = os.path.dirname(os.path.realpath(__file__))
