+++ fsleyes_plugin_shimming_toolbox/st_plugin.py
@@ -16,7 +16,6 @@ Copyright (c) 2021 Polytechnique Montreal <www.neuro.polymtl.ca>
 Authors: Alexandre D'Astous, Ainsleigh Hill, Gaspard Cereza, Julien Cohen-Adad
 """
 
-import abc
 import fsleyes.controls.controlpanel as ctrlpanel
 import fsleyes.actions.loadoverlay as loadoverlay
 import fsleyes.views.canvaspanel as canvaspanel
@@ -30,9 +29,10 @@ from pathlib import Path
 import webbrowser
 import wx
 
-from fsleyes_plugin_shimming_toolbox.events import result_event_type, EVT_RESULT, ResultEvent
-from fsleyes_plugin_shimming_toolbox.events import log_event_type, EVT_LOG, LogEvent
-from fsleyes_plugin_shimming_toolbox.worker_thread import WorkerThread
+from fsleyes_plugin_shimming_toolbox import __dir_st_plugin__
+from fsleyes_plugin_shimming_toolbox.components.dropdown import DropdownComponent
+from fsleyes_plugin_shimming_toolbox.components.run import RunComponent
+from fsleyes_plugin_shimming_toolbox.components.input import InputComponent
 from shimmingtoolbox.cli.b0shim import dynamic as dynamic_cli
 from shimmingtoolbox.cli.b0shim import realtime_dynamic as realtime_cli
 from shimmingtoolbox.cli.b0shim import max_intensity as max_intensity_cli
@@ -46,17 +46,13 @@ logger = logging.getLogger(__name__)
 HOME_DIR = str(Path.home())
 CURR_DIR = os.getcwd()
 ST_DIR = f"{HOME_DIR}/shimming-toolbox"
-DIR = os.path.dirname(__file__)
 
 VERSION = "0.1.1"
 
 # Load icon resources
-asterisk_icon = wx.Image(os.path.join(DIR, 'img', 'asterisk.png'), wx.BITMAP_TYPE_PNG).ConvertToBitmap()
-info_icon = wx.Image(os.path.join(DIR, 'img', 'info-icon.png'), wx.BITMAP_TYPE_PNG).ConvertToBitmap()
-play_icon = wx.Bitmap(os.path.join(DIR, 'img', 'play.png'), wx.BITMAP_TYPE_PNG)
-rtd_logo = wx.Bitmap(os.path.join(DIR, 'img', 'RTD.png'), wx.BITMAP_TYPE_PNG)
+rtd_logo = wx.Bitmap(os.path.join(__dir_st_plugin__, 'img', 'RTD.png'), wx.BITMAP_TYPE_PNG)
 # Load ShimmingToolbox logo saved as a png image, rescale it, and return it as a wx.Bitmap image.
-st_logo = wx.Image(os.path.join(DIR, 'img', 'shimming_toolbox_logo.png'), wx.BITMAP_TYPE_PNG)
+st_logo = wx.Image(os.path.join(__dir_st_plugin__, 'img', 'shimming_toolbox_logo.png'), wx.BITMAP_TYPE_PNG)
 st_logo.Rescale(int(st_logo.GetWidth() * 0.2), int(st_logo.GetHeight() * 0.2), wx.IMAGE_QUALITY_HIGH)
 st_logo = st_logo.ConvertToBitmap()
 
@@ -164,21 +160,6 @@ class Terminal:
             self.terminal.AppendText(f"{level}: {msg}\n")
 
 
-class Component:
-    def __init__(self, panel, list_components=[]):
-        self.panel = panel
-        self.list_components = list_components
-
-    @abc.abstractmethod
-    def create_sizer(self):
-        raise NotImplementedError
-
-    # make sure that the create_sizer method has been implemented in the subclasses
-    @classmethod
-    def __subclasshook__(cls, subclass):
-        return hasattr(subclass, 'create_sizer') and callable(subclass.create_sizer) or NotImplemented
-
-
 class InfoSection:
     def __init__(self, panel, description):
         self.panel = panel
@@ -209,498 +190,6 @@ class InfoSection:
     def open_documentation_url(self, event):
         """Redirect ``button_documentation`` to the ``shimming-toolbox`` page."""
         webbrowser.open('https://shimming-toolbox.org/en/latest/')
-
-
-class InputComponent(Component):
-    """ Define cli to automatically generate help text """
-    def __init__(self, panel, input_text_box_metadata, cli=None):
-        super().__init__(panel)
-        self.sizer = self.create_sizer()
-        self.input_text_boxes = {}
-        self.input_text_box_metadata = input_text_box_metadata
-        self.cli = cli
-        self.add_text_info()
-        self.add_input_text_boxes()
-
-    def create_sizer(self):
-        """Create the centre sizer containing tab-specific functionality."""
-        sizer = wx.BoxSizer(wx.VERTICAL)
-        return sizer
-
-    def add_text_info(self):
-        """ This function adds the help text to the metadata. This function
-            needs to be called before creating the buttons.
-
-        """
-        for i, twb_dict in enumerate(self.input_text_box_metadata):
-            if not ('info_text' in twb_dict.keys()) and (self.cli is not None) and ('name' in twb_dict.keys()):
-                description = get_help_text(self.cli, twb_dict['name'])
-                self.input_text_box_metadata[i]['info_text'] = description
-
-
-    def add_input_text_boxes(self, spacer_size=10):
-        """Add a list of input text boxes (TextWithButton) to the sizer_input.
-
-        Args:
-            self.input_text_box_metadata (list)(dict): A list of dictionaries, where the dictionaries have two keys:
-                ``button_label`` and ``button_function``.
-                .. code::
-
-                    {
-                        "button_label": The label to go on the button.
-                        "button_function": the class function (self.myfunc) which will get
-                            called when the button is pressed. If no action is desired, create
-                            a function that is just ``pass``.
-                        "default_text": (optional) The default text to be displayed.
-                        "name" : Option name in the CLI, use "arg" as the name for an argument.
-                    }
-
-            spacer_size (int): The size of the space to be placed between each input text box.
-
-        """
-        for twb_dict in self.input_text_box_metadata:
-            text_with_button = TextWithButton(
-                panel=self.panel,
-                button_label=twb_dict["button_label"],
-                button_function=twb_dict.get("button_function", self.button_do_something),
-                default_text=twb_dict.get("default_text", ""),
-                n_text_boxes=twb_dict.get("n_text_boxes", 1),
-                name=twb_dict.get("name", "default"),
-                info_text=twb_dict.get("info_text", ""),
-                required=twb_dict.get("required", False),
-                load_in_overlay=twb_dict.get("load_in_overlay", False)
-            )
-            self.add_input_text_box(text_with_button, twb_dict.get("name", "default"))
-
-    def add_input_text_box(self, text_with_button, name, spacer_size=10):
-        box = text_with_button.create()
-        self.sizer.Add(box, 1, wx.EXPAND)
-        self.sizer.AddSpacer(spacer_size)
-        if name in self.input_text_boxes.keys():
-            self.input_text_boxes[name].append(text_with_button)
-        else:
-            self.input_text_boxes[name] = [text_with_button]
-
-    def insert_input_text_box(self, text_with_button, name, index, last=False, spacer_size=10):
-        box = text_with_button.create()
-        self.sizer.Insert(index=index, sizer=box, flag=wx.EXPAND)
-        if last:
-            self.sizer.InsertSpacer(index=index + 1, size=spacer_size)
-        if name in self.input_text_boxes.keys():
-            self.input_text_boxes[name].append(text_with_button)
-        else:
-            self.input_text_boxes[name] = [text_with_button]
-
-    def remove_last_input_text_box(self, name):
-        self.input_text_boxes[name].pop(-1)
-
-    def button_do_something(self, event):
-        pass
-
-
-class DropdownComponent(Component):
-    def __init__(self, panel, dropdown_metadata, label, option_name, list_components=[], info_text=None, cli=None,
-                 component_to_dropdown_choice=None):
-        """ Create a dropdown list
-
-        Args:
-            panel (wx.Panel): A panel is a window on which controls are placed.
-            dropdown_metadata (list)(dict): A list of dictionaries where the dictionaries have the
-                required keys: ``label``, ``option_value``.
-                .. code::
-
-                    {
-                        "label": The label for the dropdown box
-                        "option_value": The value linked to the option in the CLI
-                    }
-
-            label (str): Label of the button describing the dropdown
-            option_name (str): Name of the options of the dropdown, set to 'no_arg' is not an option
-            list_components (list): list of Components
-            info_text (str): Info message displayed when hovering over the "i" icon. Leave blank to auto fill using option_name
-            cli (function): CLI function used by the dropdown
-            component_to_dropdown_choice (list): Tells which component associates with which dropdown selection. 
-                                                 If None, assumes 1:1. 
-        """
-        super().__init__(panel, list_components)
-        self.dropdown_metadata = dropdown_metadata
-        self.label = label
-        self.info_text = info_text
-        self.positions = {}
-        self.input_text_boxes = {}
-
-        # If there is a dropdown parent, it needs to be added after instanciantion with self.add_dropdown_parent(parent)
-        self.dropdown_parent = None
-
-        self.dropdown_children = self.get_dropdown_children()
-        self.sizer = self.create_sizer()
-        self.dropdown_choices = [item["label"] for item in self.dropdown_metadata]
-        self.option_name = option_name
-        self.cli = cli
-        
-        if component_to_dropdown_choice is None:
-            self.component_to_dropdown_choice = range(len(self.list_components))
-        else:
-            self.component_to_dropdown_choice = component_to_dropdown_choice
-            
-        self.add_text_info()
-        self.create_choice_box()
-        self.create_dropdown_sizers()
-        self.on_choice(None)
-
-    def get_dropdown_children(self):
-        """ Finds the DropdownComponent from list_components and adds it to the list of children (dropdown_children)"""
-        dropdown_children = []
-        for component in self.list_components:
-            if type(component) == DropdownComponent:
-                dropdown_children.append(component)
-        return dropdown_children
-
-    def add_dropdown_parent(self, parent):
-        """ Method to add the parent dropdown after instanciation of the class"""
-        self.dropdown_parent = parent
-
-    def add_text_info(self):
-        """ This function adds the help text. """
-
-        if self.info_text is None:
-            if self.cli is not None:
-                description = get_help_text(self.cli, self.option_name)
-                self.info_text = description
-            else:
-                self.info_text = ""
-
-    def create_dropdown_sizers(self):
-        for index in range(len(self.list_components)):
-            sizer = self.list_components[index].sizer
-            self.sizer.Add(sizer, 0, wx.EXPAND)
-            
-            # Map list index to the dropdown selection
-            dd_index = self.component_to_dropdown_choice[index]
-            if self.dropdown_choices[dd_index] not in self.positions.keys():
-                self.positions[self.dropdown_choices[dd_index]] = []
-            self.positions[self.dropdown_choices[dd_index]].append(self.sizer.GetItemCount() - 1)
-
-    def unshow_choice_box_sizers(self):
-        """Set the Show variable to false for all sizers of the choice box widget"""
-        for list_position in self.positions.values():
-            for position in list_position:
-                sizer = self.sizer.GetItem(position)
-                sizer.Show(False)
-
-    def create_choice_box(self):
-        self.choice_box = wx.Choice(self.panel, choices=self.dropdown_choices)
-        self.choice_box.Bind(wx.EVT_CHOICE, self.on_choice)
-        button = wx.Button(self.panel, -1, label=self.label)
-        self.choice_box_sizer = wx.BoxSizer(wx.HORIZONTAL)
-        self.choice_box_sizer.Add(create_info_icon(self.panel, self.info_text), 0, wx.ALIGN_LEFT | wx.RIGHT, 7)
-        self.choice_box_sizer.Add(button, 0, wx.ALIGN_LEFT | wx.RIGHT, 10)
-        self.choice_box_sizer.Add(self.choice_box)
-        self.sizer.Add(self.choice_box_sizer)
-        self.sizer.AddSpacer(10)
-
-    def on_choice(self, event, is_propagating_up=True):
-        """ To enable nested dropdowns, every time a user selects a dropdown, we need to recalculate all other
-            dropdowns and associated options. Dropdowns affect other dropdowns.
-
-            is_propagating_up tells the dropdown to propagate the on_choice() command to the parent drop down until
-            there is no more parent. Moreover, on each call of on_choice(), the propagation will also go down. This
-            allows to recalculate each dropdown. THere is some redundency since we would really just want the most
-            parent dropdown to send the down propagation, but this will do for now.
-        """
-        # Get the selection from the choice box widget
-        if self.choice_box.GetSelection() < 0:
-            selection = self.choice_box.GetString(0)
-            self.choice_box.SetSelection(0)
-        else:
-            selection = self.choice_box.GetString(self.choice_box.GetSelection())
-
-        # Unshow everything then show the correct item according to the choice box
-        self.unshow_choice_box_sizers()
-        if selection in self.positions.keys():
-            for a_index in self.positions[selection]:
-                sizer_item = self.sizer.GetItem(a_index)
-                sizer_item.Show(True)
-        else:
-            pass
-
-        index_dd = self.find_index(selection)
-        if selection in self.positions.keys():
-            # Add the sizers to the current list of options
-            # find indexes of list_components that are associated with index dropdown
-            indexes = [i for i, e in enumerate(self.component_to_dropdown_choice) if e == index_dd]
-            self.input_text_boxes = {}
-            for index_comp in indexes:
-                component = self.list_components[index_comp]
-                # Merge both input_text_boxes
-                self.input_text_boxes.update(component.input_text_boxes)
-                # Propagate down to show or hide the relevant sub DropdownComponents
-                if type(component) == DropdownComponent:
-                    component.on_choice(None, is_propagating_up=False)
-
-        # Add the dropdown to the list of options
-        self.input_text_boxes[self.option_name] = [self.dropdown_metadata[index_dd]["option_value"]]
-
-        # Update the parent if there is one (This allows to propagate nested dropdown selection)
-        if is_propagating_up:
-            if self.dropdown_parent is not None:
-                self.dropdown_parent.on_choice(None)
-
-        # Update the window
-        self.panel.SetVirtualSize(self.panel.sizer_run.GetMinSize())
-        self.panel.Layout()
-
-    def find_index(self, label):
-        for index in range(len(self.dropdown_metadata)):
-            if self.dropdown_metadata[index]["label"] == label:
-                return index
-
-        # Return index 0 if it is not found
-        return 0
-
-    def create_sizer(self):
-        """Create the a sizer containing tab-specific functionality."""
-        sizer = wx.BoxSizer(wx.VERTICAL)
-        return sizer
-
-
-class RunComponent(Component):
-    """Component which contains input and run button.
-
-    Attributes:
-        panel (wx.Panel): Panel, this is usually a Tab instance
-        st_function (str): Name of the ``Shimming Toolbox`` CLI function to be called.
-        list_components (list of Component): list of subcomponents to be added.
-        output_paths (list of str): file or folder paths containing output from ``st_function``.
-        output_paths (list of str): relative path of files containing output from ``st_function``. Path is relative to
-                                    the output option's folder.
-    """
-
-    def __init__(self, panel, st_function, list_components=[], output_paths=[]):
-        super().__init__(panel, list_components)
-        self.st_function = st_function
-        self.sizer = self.create_sizer()
-        self.add_button_run()
-        self.output = ""
-        self.output_paths_original = output_paths
-        self.output_paths = output_paths.copy()
-        self.load_in_overlay = []
-        self.worker = None
-
-        self.panel.Bind(EVT_RESULT, self.on_result)
-        self.panel.Bind(EVT_LOG, self.log)
-
-    def create_sizer(self):
-        """Create the centre sizer containing tab-specific functionality."""
-        sizer = wx.BoxSizer(wx.VERTICAL)
-        sizer.SetMinSize(400, 300)
-        sizer.AddSpacer(10)
-        for component in self.list_components:
-            sizer.Add(component.sizer, 0, wx.EXPAND)
-        return sizer
-
-    def add_button_run(self):
-        """Add the run button which will call the ``Shimming Toolbox`` CLI."""
-        button_run = wx.Button(self.panel, -1, label="Run", size=(85, 48))
-        button_run.Bind(wx.EVT_BUTTON, self.button_run_on_click)
-        button_run.SetBitmap(play_icon, dir=wx.LEFT)
-        self.sizer.Add(button_run, 0, wx.CENTRE)
-        self.sizer.AddSpacer(10)
-
-    def log(self, event):
-        """Log to the terminal the when there is a log event"""
-
-        # Since the log events get broadcated to all the RunComponents of the Tab, we check that the event name
-        # corresponds to our function i.e self.st_function
-        if event.name == self.st_function:
-            msg = event.get_data()
-            self.panel.terminal_component.log_to_terminal(msg)
-        else:
-            event.Skip()
-
-    def on_result(self, event):
-        # Since the log events get broadcated to all the RunComponents of the Tab, we check that the event name
-        # corresponds to our function i.e self.st_function
-        if event.name != self.st_function:
-            event.Skip()
-            return
-
-        data = event.get_data()
-        # Return code is 0 if everything ran smoothly
-        if data == 0:
-            msg = f"Run {self.st_function} completed successfully\n"
-            self.panel.terminal_component.log_to_terminal(msg, level="INFO")
-
-            # Get the directory of the output if it is a file or already a directory
-            if os.path.isfile(self.output):
-                folder = os.path.split(self.output)[0]
-            else:
-                folder = self.output
-
-            # Add the directory to the relative path of path output
-            for i_file in range(len(self.output_paths)):
-                self.output_paths[i_file] = os.path.join(folder, self.output_paths[i_file])
-
-            # Append the file if it was a file
-            if os.path.isfile(self.output):
-                self.output_paths.append(self.output)
-
-            # Append files that are a direct output from "load_in_overlay"
-            for fname in self.load_in_overlay:
-                self.output_paths.append(fname)
-
-            if self.st_function == "st_dicom_to_nifti":
-                # If its dicom_to_nifti, output all .nii found in the subject folder to the overlay
-                try:
-                    path_output, subject = self.fetch_paths_dicom_to_nifti()
-                    path_sub = os.path.join(path_output, 'sub-' + subject)
-                    list_files = sorted(glob.glob(os.path.join(path_sub, '*', '*.nii*')))
-                    for file in list_files:
-                        self.output_paths.append(file)
-
-                except Exception:
-                    self.panel.terminal_component.log_to_terminal(
-                        "Could not fetch subject and/or path to load to overlay"
-                    )
-            self.send_output_to_overlay()
-
-            self.output_paths.clear()
-            self.output_paths = self.output_paths_original.copy()
-
-        elif type(data) == Exception:
-            msg = f"Run {self.st_function} errored out\n"
-            err = data
-
-            self.panel.terminal_component.log_to_terminal(msg, level="ERROR")
-            if len(err.args) == 1:
-                # Pretty output
-                a_string = ""
-                for i_err in range(len(err.args[0])):
-                    a_string += str(err.args[0][i_err])
-                self.panel.terminal_component.log_to_terminal(a_string, level="ERROR")
-
-            else:
-                self.panel.terminal_component.log_to_terminal(str(err), level="ERROR")
-
-        else:
-            # The error message should already be displayed
-            self.panel.terminal_component.log_to_terminal(str(data), level="ERROR")
-
-        self.worker = None
-        self.load_in_overlay = []
-        event.Skip()
-
-    def button_run_on_click(self, event):
-        """Function called when the ``Run`` button is clicked.
-
-        Calls the relevant ``Shimming Toolbox`` CLI command (``st_function``) in a thread
-
-        """
-        if not self.worker:
-            try:
-                command, msg = self.get_run_args(self.st_function)
-            except RunArgumentErrorST as err:
-                self.panel.terminal_component.log_to_terminal(err, level="ERROR")
-                return
-
-            self.panel.terminal_component.log_to_terminal(msg, level="INFO")
-            self.worker = WorkerThread(self.panel, command, name=self.st_function)
-
-    def send_output_to_overlay(self):
-        for output_path in self.output_paths:
-            if os.path.isfile(output_path):
-                try:
-                    # Display the overlay
-                    window = self.panel.GetGrandParent()
-                    if output_path[-4:] == ".png":
-                        load_png_image_from_path(window, output_path, colormap="greyscale")
-                    elif output_path[-7:] == ".nii.gz" or output_path[-4:] == ".nii":
-                        # Load the NIfTI image as an overlay
-                        img_overlay = loadoverlay.loadOverlays(
-                            paths=[output_path],
-                            inmem=True,
-                            blocking=True)[0]
-                        window.overlayList.append(img_overlay)
-                except Exception as err:
-                    self.panel.terminal_component.log_to_terminal(str(err), level="ERROR")
-
-    def get_run_args(self, st_function):
-        """The option are a list of tuples where the tuple: (name, [value1, value2])"""
-        msg = "Running "
-        # Split is necessary if we have grouped commands (st_mask threshold)
-        command = st_function.split(' ')
-
-        self.output = ""
-        command_list_arguments = []
-        command_list_options = []
-        for component in self.list_components:
-            for name, input_text_box_list in component.input_text_boxes.items():
-
-                if name == "no_arg":
-                    continue
-
-                for input_text_box in input_text_box_list:
-                    # Allows to choose from a dropdown
-                    if type(input_text_box) == str:
-                        command_list_options.append((name, [input_text_box]))
-
-                    # Normal case where input_text_box is a TextwithButton
-                    else:
-                        is_arg = False
-                        option_values = []
-                        for textctrl in input_text_box.textctrl_list:
-                            arg = textctrl.GetValue()
-                            if arg == "" or arg is None:
-                                if input_text_box.required is True:
-                                    raise RunArgumentErrorST(
-                                        f"Argument {name} is missing a value, please enter a valid input"
-                                    )
-                            else:
-                                # Case where the option name is set to arg, this handles it as if it were an argument
-                                if name == "arg":
-                                    command_list_arguments.append(arg)
-                                    is_arg = True
-                                # Normal options
-                                else:
-                                    if name == "output":
-                                        self.output = arg
-                                    elif input_text_box.load_in_overlay:
-                                        self.load_in_overlay.append(arg)
-
-                                    option_values.append(arg)
-
-                        # If its an argument don't include it as an option, if the option list is empty don't either
-                        if not is_arg and option_values:
-                            command_list_options.append((name, option_values))
-
-        # Arguments don't need "-"
-        for arg in command_list_arguments:
-            command += [arg]
-
-        # Handles options
-        for name, args in command_list_options:
-            command += ['--' + name]
-            for arg in args:
-                command += [arg]
-
-        msg += ' '.join(command) + '\n'
-
-        return command, msg
-
-    def fetch_paths_dicom_to_nifti(self):
-        if self.st_function == "st_dicom_to_nifti":
-
-            for component in self.list_components:
-
-                if 'subject' in component.input_text_boxes.keys():
-                    box_with_button_sub = component.input_text_boxes['subject'][0]
-                    subject = box_with_button_sub.textctrl_list[0].GetValue()
-
-                if 'output' in component.input_text_boxes.keys():
-                    box_with_button_out = component.input_text_boxes['output'][0]
-                    path_output = box_with_button_out.textctrl_list[0].GetValue()
-
-            return path_output, subject
 
 
 class B0ShimTab(Tab):
@@ -2111,371 +1600,6 @@ class DicomToNiftiTab(Tab):
         return run_component.sizer
 
 
-class TextWithButton:
-    """Creates a button with an input text box.
-
-    wx.BoxSizer:
-
-        InfoIcon(wx.StaticBitmap) - info icon
-        wx.Button - clickable input button
-        [wx.TextCtrl] - input text box(es)
-        wx.StaticBitmap - asterisk icon
-
-    Attributes:
-
-        panel (wx.Panel): Instance of a Panel, this is usually a Tab
-        button_label (str): label to be put on the button.
-        button_function: function which gets called when the button is clicked on. If it's a list, assign to the
-                         'n_text_boxes'
-        default_text (str): (optional) default text to be displayed in the input text box.
-        textctrl_list (list wx.TextCtrl): list of input text boxes, can be more than one in a row.
-        n_text_boxes (int): number of input text boxes to create.
-        name (str): name of the cli option
-        info_text (str): text to be displayed when hovering over the info icon; should describe
-            what the button/input is for.
-        required (bool): if this input is required or not. If True, a red asterisk will be
-            placed next to the input text box to indicate this.
-    """
-
-    def __init__(self, panel, button_label, button_function, name="default", default_text="",
-                 n_text_boxes=1, info_text="", required=False, load_in_overlay=False):
-        self.panel = panel
-        self.button_label = button_label
-        if type(button_function) is not list:
-            button_function = [button_function]
-        self.button_function = button_function
-        self.default_text = default_text
-        self.textctrl_list = []
-        self.n_text_boxes = n_text_boxes
-        self.name = name
-        self.info_text = info_text
-        self.required = required
-        self.load_in_overlay = load_in_overlay
-
-    def create(self):
-        text_with_button_box = wx.BoxSizer(wx.HORIZONTAL)
-        button = wx.Button(self.panel, -1, label=self.button_label)
-        text_with_button_box.Add(
-            create_info_icon(self.panel, self.info_text), 0, wx.ALIGN_LEFT | wx.RIGHT, 7
-        )
-        for i_text_box in range(0, self.n_text_boxes):
-            textctrl = wx.TextCtrl(parent=self.panel, value=self.default_text, name=self.name)
-            self.textctrl_list.append(textctrl)
-
-        if len(self.button_function) > self.n_text_boxes:
-            raise RuntimeError("button_function has more functions than the number of input boxes")
-
-        if len(self.button_function) == 1:
-            focus = False
-        else:
-            focus = True
-
-        for i, button_function in enumerate(self.button_function):
-            if button_function == "select_folder":
-                function = lambda event, panel=self.panel, ctrl=self.textctrl_list[i]: \
-                    select_folder(event, panel, ctrl, focus)
-                button.Bind(wx.EVT_BUTTON, function)
-            elif button_function == "select_file":
-                function = lambda event, panel=self.panel, ctrl=self.textctrl_list[i]: \
-                    select_file(event, panel, ctrl, focus)
-                button.Bind(wx.EVT_BUTTON, function)
-            elif button_function == "select_from_overlay":
-                function = lambda event, panel=self.panel, ctrl=self.textctrl_list[i]: \
-                    select_from_overlay(event, panel, ctrl, focus)
-                button.Bind(wx.EVT_BUTTON, function)
-            elif button_function == "add_input_phase_boxes":
-                function = lambda event, panel=self.panel, ctrl=self.textctrl_list[i]: \
-                    add_input_phase_boxes(event, panel, ctrl)
-                self.textctrl_list[i].Bind(wx.EVT_TEXT, function)
-            elif button_function == "add_input_coil_boxes_dyn":
-                function = lambda event, panel=self.panel, ctrl=self.textctrl_list[i], index=1: \
-                    add_input_coil_boxes(event, panel, ctrl, index)
-                self.textctrl_list[i].Bind(wx.EVT_TEXT, function)
-            elif button_function == "add_input_coil_boxes_rt":
-                function = lambda event, panel=self.panel, ctrl=self.textctrl_list[i], index=2: \
-                    add_input_coil_boxes(event, panel, ctrl, index)
-                self.textctrl_list[i].Bind(wx.EVT_TEXT, function)
-
-        text_with_button_box.Add(button, 0, wx.ALIGN_LEFT | wx.RIGHT, 10)
-
-        for textctrl in self.textctrl_list:
-            text_with_button_box.Add(textctrl, 1, wx.ALIGN_LEFT | wx.LEFT, 10)
-            if self.required:
-                text_with_button_box.Add(wx.StaticBitmap(self.panel, bitmap=asterisk_icon), 0, wx.RIGHT, 7)
-
-        return text_with_button_box
-
-
-def create_info_icon(panel, info_text=""):
-    image = InfoIcon(panel, bitmap=info_icon, info_text=info_text)
-    image.Bind(wx.EVT_MOTION, on_info_icon_mouse_over)
-    return image
-
-
-def on_info_icon_mouse_over(event):
-    image = event.GetEventObject()
-    image.SetToolTip(wx.ToolTip(image.info_text))
-
-
-class InfoIcon(wx.StaticBitmap):
-    def __init__(self, panel, bitmap, info_text):
-        self.info_text = info_text
-        super(wx.StaticBitmap, self).__init__(panel, bitmap=bitmap)
-
-
-def select_folder(event, tab, ctrl, focus=False):
-    """Select a file folder from system path."""
-    if focus:
-        # Skip allows to handle other events
-        focused = wx.Window.FindFocus()
-        if ctrl != focused:
-            if focused == tab:
-                tab.terminal_component.log_to_terminal("Select a text box from the same row.", level="INFO")
-                # If its the tab, don't handle the other events so that the message is only logged once
-                return
-            event.Skip()
-            return
-
-    dlg = wx.DirDialog(None, "Choose Directory", CURR_DIR, wx.DD_DEFAULT_STYLE | wx.DD_DIR_MUST_EXIST)
-
-    if dlg.ShowModal() == wx.ID_OK:
-        folder = dlg.GetPath()
-        ctrl.SetValue(folder)
-        logger.info(f"Folder set to: {folder}")
-
-    # Skip allows to handle other events
-    event.Skip()
-
-
-def select_file(event, tab, ctrl, focus=False):
-    """Select a file from system path."""
-    if focus:
-        # Skip allows to handle other events
-        focused = wx.Window.FindFocus()
-        if ctrl != focused:
-            if focused == tab:
-                tab.terminal_component.log_to_terminal("Select a text box from the same row.", level="INFO")
-                # If it's the tab, don't handle the other events so that the log message is only displayed once
-                return
-            event.Skip()
-            return
-
-    dlg = wx.FileDialog(parent=None,
-                        message="Select File",
-                        defaultDir=CURR_DIR,
-                        style=wx.FD_DEFAULT_STYLE | wx.FD_FILE_MUST_EXIST)
-
-    if dlg.ShowModal() == wx.ID_OK:
-        fname = dlg.GetPath()
-        ctrl.SetValue(fname)
-        logger.info(f"File set to: {fname}")
-
-    # Skip allows to handle other events
-    event.Skip()
-
-
-def select_from_overlay(event, tab, ctrl, focus=False):
-    """Fetch path to file highlighted in the Overlay list.
-
-    Args:
-        event (wx.Event): event passed to a callback or member function.
-        tab (Tab): Must be a subclass of the Tab class
-        ctrl (wx.TextCtrl): the text item.
-        focus (bool): Tells whether the ctrl must be in focus.
-    """
-    if focus:
-        # Skip allows to handle other events
-        focused = wx.Window.FindFocus()
-        if ctrl != focused:
-            if focused == tab:
-                tab.terminal_component.log_to_terminal(
-                    "Select a text box from the same row.",
-                    level="INFO"
-                )
-                # If its the tab, don't handle the other events so that the log message is only once
-                return
-            event.Skip()
-            return
-
-    # This is messy and wont work if we change any class hierarchy.. using GetTopLevelParent() only
-    # works if the pane is not floating
-    # Get the displayCtx class initialized in STControlPanel
-    window = tab.GetGrandParent()
-    selected_overlay = window.displayCtx.getSelectedOverlay()
-    if selected_overlay is not None:
-        filename_path = selected_overlay.dataSource
-        ctrl.SetValue(filename_path)
-    else:
-        tab.terminal_component.log_to_terminal(
-            "Import and select an image from the Overlay list",
-            level="INFO"
-        )
-
-    # Skip allows to handle other events
-    event.Skip()
-
-
-def add_input_phase_boxes(event, tab, ctrl):
-    """On click of ``Number of Echoes`` button, add ``n_echoes`` ``TextWithButton`` boxes.
-
-    For this function, we are assuming the layout of the Component input is as follows:
-
-        0 - Number of Echoes TextWithButton sizer
-        1 - Spacer
-        2 - next item, and so on
-
-    First, we check and see how many phase boxes the tab currently has, and remove any where
-    n current > n update.
-    Next, we add n = n update - n current phase boxes to the tab.
-
-    Args:
-        event (wx.Event): when the ``Number of Echoes`` button is clicked.
-        tab (FieldMapTab): tab class instance for ``Field Map``.
-        ctrl (wx.TextCtrl): the text box containing the number of phase boxes to add. Must be an
-            integer > 0.
-    """
-    option_name = "arg"
-    try:
-        n_echoes = int(ctrl.GetValue())
-        if n_echoes < 1:
-            raise Exception()
-        elif n_echoes > 6:
-            n_echoes = 6
-            tab.terminal_component.log_to_terminal("Number of echoes limited to 6", level="WARNING")
-    except Exception:
-        tab.terminal_component.log_to_terminal("Number of Echoes must be an integer > 0", level="ERROR")
-        return
-
-    insert_index = 2
-    if n_echoes < tab.n_echoes:
-        for index in range(tab.n_echoes, n_echoes, - 1):
-            tab.component_input.sizer.Hide(index + 1)
-            tab.component_input.sizer.Remove(index + 1)
-            tab.component_input.remove_last_input_text_box(option_name)
-
-    for index in range(tab.n_echoes, n_echoes):
-        text_with_button = TextWithButton(
-            panel=tab,
-            button_label=f"Input Phase {index + 1}",
-            button_function="select_from_overlay",
-            default_text="",
-            n_text_boxes=1,
-            name=f"input_phase_{index + 1}",
-            info_text=f"Input path of phase nifti file {index + 1}",
-            required=True
-        )
-        if index + 1 == n_echoes and tab.n_echoes == 0:
-            tab.component_input.insert_input_text_box(
-                text_with_button,
-                option_name,
-                index=insert_index + index,
-                last=True)
-        else:
-            tab.component_input.insert_input_text_box(
-                text_with_button,
-                option_name,
-                index=insert_index + index
-            )
-
-    tab.n_echoes = n_echoes
-    tab.SetVirtualSize(tab.sizer_run.GetMinSize())
-
-
-def add_input_coil_boxes(event, tab, ctrl, i=0):
-    """On click of ``Number of Custom Coils`` button, add ``n_coils`` ``TextWithButton`` boxes.
-
-    For this function, we are assuming the layout of the Component input is as follows:
-
-        0 - Number of Coils TextWithButton sizer
-        1 - Spacer
-        2 - next item, and so on
-
-    First, we check and see how many coil boxes the tab currently has, and remove any where
-    n current > n update.
-    Next, we add n = n update - n current coil boxes to the tab.
-
-    Args:
-        event (wx.Event): when the ``Number of Echoes`` button is clicked.
-        tab (B0ShimTab): tab class instance for ``B0 Shim``.
-        ctrl (wx.TextCtrl): the text box containing the number of phase boxes to add. Must be an
-            integer > 0.
-        i (int): Index of the coil instance. Used when the tab has multiple coil instances. 1 <= index <= 2
-    """
-
-    option_name = "coil"
-    try:
-        if ctrl.GetValue() == "":
-            n_coils = 0
-        else:
-            n_coils = int(ctrl.GetValue())
-        if n_coils < 0:
-            raise Exception()
-        elif n_coils > 5:
-            n_coils = 5
-            tab.terminal_component.log_to_terminal("Number of coils limited to 5", level="WARNING")
-
-    except Exception:
-        tab.terminal_component.log_to_terminal("Number of coils must be an integer >= 0", level="ERROR")
-        n_coils = 0
-
-    # Depending on the index, select the appropriate component
-    if i == 1:
-        n_coils_displayed = tab.n_coils_dyn
-        component_coils = tab.component_coils_dyn
-    elif i == 2:
-        n_coils_displayed = tab.n_coils_rt
-        component_coils = tab.component_coils_rt
-    else:
-        raise NotImplementedError("Index of the coil instance not implemented for more indexes")
-
-    insert_index = 2
-    # If we have to remove coils
-    if n_coils < n_coils_displayed:
-        for index in range(n_coils_displayed, n_coils, - 1):
-            component_coils.sizer.Hide(index + 1)
-            component_coils.sizer.Remove(index + 1)
-            component_coils.remove_last_input_text_box(option_name)
-
-        # Delete the last spacer if we go back to n_coils == 0
-        if n_coils == 0:
-            index = 2
-            component_coils.sizer.Hide(index)
-            component_coils.sizer.Remove(index)
-
-    for index in range(n_coils_displayed, n_coils):
-        text_with_button = TextWithButton(
-            panel=tab,
-            button_label=f"Input Coil {index + 1}",
-            button_function=["select_from_overlay", "select_file"],
-            default_text="",
-            n_text_boxes=2,
-            name=f"input_coil_{index + 1}",
-            info_text=f"Input path of the coil nifti file followed by json constraint file",
-            required=True
-        )
-        # Add a spacer at the end if its the last one and if there were none previously
-        # i.e. if it was previously n_coils == 0
-        if index + 1 == n_coils and n_coils_displayed == 0:
-            component_coils.insert_input_text_box(
-                text_with_button,
-                option_name,
-                index=insert_index + index,
-                last=True)
-        else:
-            component_coils.insert_input_text_box(
-                text_with_button,
-                option_name,
-                index=insert_index + index
-            )
-
-    if i == 1:
-        tab.n_coils_dyn = n_coils
-    elif i == 2:
-        tab.n_coils_rt = n_coils
-
-    tab.SetVirtualSize(tab.sizer_run.GetMinSize())
-
-
 def read_image(filename, bitdepth=8):
     """Read image and convert it to desired bitdepth without truncation."""
     if 'tif' in str(filename):
@@ -2545,19 +1669,3 @@ def load_png_image_from_path(fsl_panel, image_path, is_mask=False, add_to_overla
         opts.cmap = colormap
 
     return img_overlay
-
-
-def get_help_text(cli_function, name):
-    """ Returns the help text of a cli function depending on its name. """
-    for param in cli_function.params:
-        # Try different versions of the input dashes
-        for dashes in ['', '-', '--']:
-            new_name = dashes + name
-            if (new_name in param.opts) or new_name == param.human_readable_name:
-                return param.help
-
-    raise ValueError(f"Could not find param: {name} in {cli_function.name}")
-
-class RunArgumentErrorST(Exception):
-    """Exception for missing input arguments for CLI call."""
-    pass
